pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mt5-ai-bot"
version = "0.1.0"
description = "Neural market prediction stack for MetaTrader 5"
authors = [{name = "AI Agents"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "torch>=2.2.0",
  "scikit-learn>=1.4.0",
  "pyyaml>=6.0.1",
  "joblib>=1.3.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
